fix smalldata part file clean up for non str filename and for MPI

+++ psana/psana/smalldata.py
@@ -471,6 +471,7 @@ class Server: # (hdf5 handling)
                 if cache.n_events > 0:
                     self.write_to_file(dset, cache)
             self.file_handle.close()
+            time.sleep(10)
         return
 
 
@@ -530,14 +531,18 @@ class SmallData: # (client)
             print('setting cache_size -->', batch_size)
             cache_size = batch_size
 
-        self._full_filename = filename
+        self._full_filename = str(filename)
         if (filename is not None):
             self._basename = os.path.basename(filename)
             self._dirname  = os.path.dirname(filename)
 
-            # clean up previous part files associated with the filename
-            for f in glob.glob( self._full_filename.replace('.h5','_part*.h5') ):
-                os.remove(f)
+            # clean up previous part files associated with the filename.
+            # This needs to be done on a single rank to avoid trying to delete i
+            # the same file multiple times.
+            if self._type == 'client' and self._full_filename is not None:
+                if self._client_comm.Get_rank() == 0:
+                    for f in glob.glob( self._full_filename.replace('.h5','_part*.h5') ):
+                        os.remove(f)
 
         self._first_open = True # filename has not been opened yet
